Separated tests into unit & integration
