+++ tests/resources/roi2.roi
