+++ tests/unit/test_normalise_image.py
