+++ tests/unit/test_process_roi.py
@@ -1,10 +1,7 @@
 from __future__ import annotations
 
 import numpy as np
-import pytest
-from matplotlib.path import Path
 
-from cellphe.input import read_roi
 from cellphe.processing.roi import boundary_vertices, roi_corners
 
 
@@ -52,50 +49,4 @@ def test_roi_corners_diamond():
     expected = expected - input.min(axis=0)
     input = input - input.min(axis=0)
     output = roi_corners(input)
-    assert (np.sort(expected, axis=0) == np.sort(output, axis=0)).all()
-
-
-# TODO Put in integration test!
-def test_roi_corners_real_roi():
-    roi = read_roi("tests/resources/roi.roi")
-    roi = roi - roi.min(axis=0)
-    output = roi_corners(roi)
-    expected = np.array(
-        [
-            [0, 7],
-            [1, 6],
-            [1, 5],
-            [4, 2],
-            [5, 2],
-            [6, 1],
-            [7, 1],
-            [8, 0],
-            [14, 0],
-            [15, 1],
-            [16, 1],
-            [17, 2],
-            [18, 2],
-            [21, 5],
-            [21, 7],
-            [22, 8],
-            [22, 13],
-            [21, 14],
-            [21, 15],
-            [20, 16],
-            [20, 17],
-            [19, 18],
-            [18, 18],
-            [17, 19],
-            [3, 19],
-            [2, 18],
-            [2, 17],
-            [1, 16],
-            [1, 14],
-            [0, 13],
-            [1, 19],  # This and the following vertex should have been removed!
-            [1, 18],
-        ]
-    )
-    expected = np.flip(expected, axis=1)
-
     assert (np.sort(expected, axis=0) == np.sort(output, axis=0)).all()
