+++ tests/unit/test_seg_errors.py
